service/requirements.txt
google-adk>=1.2.1
google-genai>=1.17.0
fastapi>=0.124.1,<1
uvicorn[standard]==0.34.2
pydantic>=2.11.4
python-multipart==0.0.20
python-dotenv==1.1.0
httpx>=0.27,<1
loguru==0.7.3
